+++ relay-conventions/src/lib.rs
@@ -20,9 +20,6 @@ pub mod consts {
         #[deprecated(note = "Needs to be defined in sentry-conventions")]
         pub const HTTP__REQUEST_METHOD: &str = "http.request_method";
 
-        #[deprecated(note = "Needs to be defined in sentry-conventions")]
-        pub const SENTRY__WAS_TRANSACTION: &str = "sentry.was_transaction";
-
         // TODO(mjq): Evaluate and remove usages of this constant, or restore it in
         // conventions. This was deleted from conventions in
         // https://github.com/getsentry/sentry-conventions/pull/256 but is still